requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
exceptiongroup==1.0.1
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
python-decouple==3.6
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
